LayoutTests/fast/canvas/winding-enumeration-expected.txt

PASS context.fill() did not throw exception.
PASS context.fill('nonzero') did not throw exception.
PASS context.fill('randomstring') threw exception TypeError: Type error.
PASS context.clip() did not throw exception.
PASS context.clip('nonzero') did not throw exception.
PASS context.clip('randomstring') threw exception TypeError: Type error.
PASS context.isPointInPath(0, 0) did not throw exception.
PASS context.isPointInPath(0, 0, 'nonzero') did not throw exception.
PASS context.isPointInPath(0, 0, 'randomstring') threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
